+++ spec/support/vcr.rb
@@ -4,6 +4,6 @@ require 'webmock/rspec'
 VCR.configure do |config|
   config.cassette_library_dir = 'spec/fixtures/slack'
   config.hook_into :webmock
-  config.default_cassette_options = { record: :new_episodes }
+  # config.default_cassette_options = { record: :new_episodes }
   config.configure_rspec_metadata!
 end