+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.8.0 (Next)
 
-* [#134](https://github.com/slack-ruby/slack-ruby-client/issues/134): Added `timeout` and `open_timeout` options to Web API - [@dblock](https://github.com/dblock).
+* [#135](https://github.com/slack-ruby/slack-ruby-client/issues/135): Added `timeout` and `open_timeout` options to Web API - [@dblock](https://github.com/dblock).
+* [#134](https://github.com/slack-ruby/slack-ruby-client/issues/134): Set `start_options[:request][:timeout]`, used with `rtm.start` in `Slack::RealTime::Client`, to 180 seconds - [@dblock](https://github.com/dblock).
+* [#136](https://github.com/slack-ruby/slack-ruby-client/pull/136): Pass request options in web client calls - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.7.9 (2/9/2017)